+++ playwright.config.js
@@ -14,7 +14,6 @@ export default defineConfig({
     ],
     use: {
         baseURL,
-        permissions: ["clipboard-read", "clipboard-write"],
         trace: "retain-on-failure",
         screenshot: "only-on-failure",
         video: "retain-on-failure"
@@ -22,7 +21,22 @@ export default defineConfig({
     projects: [
         {
             name: "chromium",
-            use: { ...devices["Desktop Chrome"] }
+            use: {
+                ...devices["Desktop Chrome"],
+                permissions: ["clipboard-read", "clipboard-write"]
+            }
+        },
+        {
+            name: "firefox",
+            use: {
+                ...devices["Desktop Firefox"]
+            }
+        },
+        {
+            name: "webkit",
+            use: {
+                ...devices["Desktop Safari"]
+            }
         }
     ],
     webServer: {